refactor auth command into a separate module
