+++ cognite_toolkit/_cdf_tk/apps/_core_app.py
@@ -23,11 +23,11 @@ from cognite_toolkit._cdf_tk.commands import (
     DeployV2Command,
 )
 from cognite_toolkit._cdf_tk.commands._base import AVAILABLE_DATA_TYPES
+from cognite_toolkit._cdf_tk.commands.auth import EnvironmentVariables
 from cognite_toolkit._cdf_tk.commands.build_v2.data_classes import BuildParameters, ConfigYAML
 from cognite_toolkit._cdf_tk.exceptions import ToolkitFileNotFoundError
 from cognite_toolkit._cdf_tk.tk_warnings import ToolkitDeprecationWarning
 from cognite_toolkit._cdf_tk.utils import humanize_collection
-from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from cognite_toolkit._cdf_tk.utils.file import relative_to_if_possible
 from cognite_toolkit._version import __version__ as current_version
