+++ cognite_toolkit/_cdf_tk/commands/auth/__init__.py
@@ -1,5 +1,12 @@
 """Persisted CogIdP authentication for the Cognite Toolkit CLI."""
 
 from .command import AuthCommand
+from .data_classes import CLIENT_NAME, EnvironmentVariables
+from .utils import parse_login_flow_input
 
-__all__ = ["AuthCommand"]
+__all__ = [
+    "AuthCommand",
+    "CLIENT_NAME",
+    "EnvironmentVariables",
+    "parse_login_flow_input",
+]
