+++ cognite_toolkit/_cdf_tk/commands/auth/data_classes/__init__.py
@@ -1,0 +1,22 @@
+from ._constants import (
+    CLIENT_NAME,
+    LOGIN_FLOW_DESCRIPTION,
+    PROVIDER_DESCRIPTION,
+    VALID_LOGIN_FLOWS,
+    VALID_PROVIDERS,
+)
+from ._env_options import EnvOptions
+from ._environment_variables import EnvironmentVariables
+from ._types import LoginFlow, Provider
+
+__all__ = [
+    "CLIENT_NAME",
+    "LOGIN_FLOW_DESCRIPTION",
+    "PROVIDER_DESCRIPTION",
+    "VALID_LOGIN_FLOWS",
+    "VALID_PROVIDERS",
+    "EnvOptions",
+    "EnvironmentVariables",
+    "LoginFlow",
+    "Provider",
+]
