+++ cognite_toolkit/_cdf_tk/commands/auth/data_classes/_environment_variables.py
@@ -1,10 +1,7 @@
 import os
-from collections.abc import Iterable, Mapping
 from dataclasses import Field, dataclass, field, fields
-from typing import Any, Literal, TypeAlias, get_args
+from typing import Any
 
-import questionary
-import typer
 from cognite.client.credentials import (
     CredentialProvider,
     OAuthClientCredentials,
@@ -12,90 +9,16 @@ from cognite.client.credentials import (
     OAuthInteractive,
     Token,
 )
-from questionary import Choice
-from rich import print
 from rich.console import Console
 
 from cognite_toolkit._cdf_tk.client import ToolkitClient, ToolkitClientConfig
 from cognite_toolkit._cdf_tk.constants import TOOLKIT_CLIENT_ENTRA_ID
 from cognite_toolkit._cdf_tk.exceptions import AuthenticationError, ToolkitKeyError, ToolkitMissingValueError
 from cognite_toolkit._cdf_tk.utils import humanize_collection
-from cognite_toolkit._version import __version__
 
-Provider: TypeAlias = Literal["entra_id", "auth0", "cdf", "other"]
-LoginFlow: TypeAlias = Literal["client_credentials", "token", "device_code", "interactive", "session"]
-AuthLoginFlowCli: TypeAlias = Literal["session", "device-code", "client-credentials", "token", "interactive"]
-VALID_PROVIDERS = get_args(Provider)
-VALID_LOGIN_FLOWS = get_args(LoginFlow)
-VALID_AUTH_LOGIN_FLOWS = get_args(AuthLoginFlowCli)
-
-CLIENT_NAME = f"CDF-Toolkit:{__version__}"
-
-_AUTH_LOGIN_FLOW_TO_LOGIN_FLOW: dict[AuthLoginFlowCli, LoginFlow] = {
-    "session": "session",
-    "device-code": "device_code",
-    "client-credentials": "client_credentials",
-    "token": "token",
-    "interactive": "interactive",
-}
-
-
-def parse_auth_login_flow(flow: AuthLoginFlowCli) -> LoginFlow:
-    return _AUTH_LOGIN_FLOW_TO_LOGIN_FLOW[flow]
-
-
-def read_env_login_flow() -> LoginFlow | None:
-    login_flow = os.environ.get("LOGIN_FLOW", "").strip().lower()
-    if login_flow in VALID_LOGIN_FLOWS:
-        return login_flow  # type: ignore[return-value]
-    return None
-
-
-PROVIDER_DESCRIPTION = {
-    "entra_id": "Use Microsoft Entra ID to authenticate",
-    "auth0": "Use Auth0 to authenticate",
-    "cdf": "Use Cognite IDP to authenticate",
-    "other": "Use other IDP to authenticate",
-}
-LOGIN_FLOW_DESCRIPTION = {
-    "session": "Sign in via CogIdP and persist a refreshable session (recommended for local development)",
-    "device_code": "Sign in via a browser on any device — no service principal needed (best for local setup)",
-    "interactive": "Sign in via the browser on this machine with your user credentials",
-    "client_credentials": "Use a service principal with client ID and secret (for CI/CD or automated workloads)",
-    "token": "Supply a pre-existing token directly",
-}
-
-
-@dataclass
-class EnvOptions(Mapping):
-    display_name: str
-    default_example: str = ""
-    example: dict[Provider, str] = field(default_factory=dict)
-    is_secret: bool = False
-    required: frozenset[tuple[Provider | None, LoginFlow]] = frozenset()
-    optional: frozenset[tuple[Provider | None, LoginFlow]] = frozenset()
-
-    def __getitem__(self, key: str) -> str | bool:
-        return self.__dict__[key]
-
-    def __iter__(self) -> Iterable[str]:  # type: ignore[override]
-        return iter(self.__dict__.keys())
-
-    def __len__(self) -> int:
-        return len(self.__dict__)
-
-
-ALL_CASES = [(None, flow) for flow in VALID_LOGIN_FLOWS]
-
-
-def all_providers(
-    flow: LoginFlow, exclude: Provider | Iterable[Provider] | None = None
-) -> frozenset[tuple[Provider | None, LoginFlow]]:
-    if isinstance(exclude, str):
-        exclude = {exclude}
-    elif exclude is not None:
-        exclude = set(exclude)
-    return frozenset((prov, flow) for prov in VALID_PROVIDERS if exclude is None or prov not in exclude)
+from ._env_options import ALL_CASES, EnvOptions, all_providers
+from ._constants import CLIENT_NAME, VALID_LOGIN_FLOWS, VALID_PROVIDERS, parse_login_flow
+from ._types import LoginFlow, Provider
 
 
 @dataclass
@@ -203,11 +126,22 @@ class EnvironmentVariables:
     _client: ToolkitClient | None = field(init=False, default=None)
 
     def __post_init__(self) -> None:
-        self.LOGIN_FLOW = self.LOGIN_FLOW.lower()  # type: ignore[assignment]
-        if self.LOGIN_FLOW not in VALID_LOGIN_FLOWS:
-            raise AuthenticationError(f"Invalid login flow: {self.LOGIN_FLOW}. Valid options are {VALID_LOGIN_FLOWS}")
+        try:
+            self.LOGIN_FLOW = parse_login_flow(self.LOGIN_FLOW)
+        except ValueError as exc:
+            raise AuthenticationError(str(exc)) from exc
         if self.PROVIDER not in VALID_PROVIDERS:
             raise AuthenticationError(f"Invalid provider: {self.PROVIDER}. Valid options are {VALID_PROVIDERS}")
+
+    @classmethod
+    def login_flow_from_environment(cls) -> LoginFlow | None:
+        raw = os.environ.get("LOGIN_FLOW", "").strip()
+        if not raw:
+            return None
+        try:
+            return parse_login_flow(raw)
+        except ValueError:
+            return None
 
     # All derived properties
     @property
@@ -487,139 +421,3 @@ class EnvironmentVariables:
                 value = ",".join(value)
             lines.append(f"{field_.name}={value}")
         return "\n".join(lines) + "\n"
-
-
-def prompt_user_environment_variables(
-    current: EnvironmentVariables | None = None,
-    *,
-    login_flow: LoginFlow | None = None,
-) -> EnvironmentVariables:
-    provider = questionary.select(
-        "Choose the provider (Who authenticates you?)",
-        choices=[
-            Choice(title=f"{provider}: {description}", value=provider)
-            # We CDF as a provider is not yet out of alpha, so we hide it from the user.
-            for provider, description in PROVIDER_DESCRIPTION.items()
-            if provider != "cdf"
-        ],
-        default=current.PROVIDER if current else "entra_id",
-    ).unsafe_ask()
-    exclude = set()
-    if provider == "cdf":
-        exclude = set(VALID_LOGIN_FLOWS) - {"client_credentials"}
-    choices = [
-        Choice(title=f"{flow}: {description}", value=flow)
-        for flow, description in LOGIN_FLOW_DESCRIPTION.items()
-        if flow not in exclude and flow != "session"
-    ]
-    if login_flow is None:
-        if len(choices) == 1:
-            print(f"Only one login flow available: {choices[0].title}")
-            login_flow = choices[0].value
-        else:
-            login_flow = questionary.select(
-                "Choose the login flow (How do you going to authenticate?)",
-                choices=choices,
-                default=current.LOGIN_FLOW if current else "device_code",
-            ).unsafe_ask()
-
-    cdf_cluster = questionary.text(
-        "Enter the CDF cluster (e.g. westeurope-1)",
-        default=current.CDF_CLUSTER if current else "",
-        validate=lambda v: True if v.strip() else "CDF cluster cannot be empty",
-    ).unsafe_ask()
-    cdf_project = questionary.text(
-        "Enter the CDF project",
-        default=current.CDF_PROJECT if current else "",
-        validate=lambda v: True if v.strip() else "CDF project cannot be empty",
-    ).unsafe_ask()
-    args: dict[str, Any] = (
-        current.dump(include_os=False)
-        if current and _is_unchanged(current, provider, login_flow, cdf_project, cdf_cluster)  # type: ignore[arg-type]
-        else {}
-    )
-    args.update(
-        {"LOGIN_FLOW": login_flow, "CDF_CLUSTER": cdf_cluster, "CDF_PROJECT": cdf_project, "PROVIDER": provider}
-    )
-    env_vars = EnvironmentVariables(**args)
-    idp_tenant_id = env_vars.IDP_TENANT_ID or "IDP_TENANT_ID"
-    for field_, value in env_vars.get_required_with_value():
-        user_value = get_user_value(field_, value, provider, cdf_cluster, cdf_project, idp_tenant_id)
-        setattr(env_vars, field_.name, user_value)
-        if field_.name == "IDP_TENANT_ID":
-            idp_tenant_id = user_value
-
-    optional_values = env_vars.get_optional_with_value()
-    if optional_values:
-        print("Additional variables:")
-        for field_, value in optional_values:
-            print(f"  {field_.name}={value}")
-        if questionary.confirm("Do you want to change any of these variables?", default=False).unsafe_ask():
-            for field_, value in optional_values:
-                user_value = get_user_value(field_, value, provider, cdf_cluster, cdf_project, idp_tenant_id)
-                setattr(env_vars, field_.name, user_value)
-    return env_vars
-
-
-def get_user_value(
-    field_: Field, value: Any, provider: Provider, cdf_cluster: str, cdf_project: str, idp_tenant_id: str
-) -> Any:
-    is_secret = field_.metadata["is_secret"]
-    display_name = field_.metadata["display_name"]
-    hint = ""
-    if value:
-        default = value
-    else:
-        default_example = field_.metadata["default_example"]
-        example = (
-            field_.metadata["example"]
-            .get(provider, default_example)
-            .format(CDF_CLUSTER=cdf_cluster, CDF_PROJECT=cdf_project, IDP_TENANT_ID=idp_tenant_id)
-        )
-        # Use the example as a pre-filled default only when it is a concrete value
-        # (i.e. all template placeholders have been resolved).  Strings that still
-        # look like documentation hints (contain "<", ">", or " or ") are shown as
-        # hint text in the question label instead, so the user cannot accidentally
-        # submit them verbatim.
-        if example and not any(marker in example for marker in ("<", ">", " or ")):
-            default = example
-        else:
-            default = ""
-            if example:
-                hint = f" (e.g. {example})"
-
-    if isinstance(value, list):
-        default = ",".join(value)
-    elif value is not None and not isinstance(value, str):
-        default = str(value)
-    prompt = f"Enter the {display_name}{hint}:"
-    if is_secret:
-        user_value = questionary.password(prompt, default=default).unsafe_ask()
-    else:
-        user_value = questionary.text(prompt, default=default).unsafe_ask()
-    if user_value is None:
-        raise typer.Exit(0)
-    if field_.type is int:
-        try:
-            user_value = int(user_value)
-        except ValueError:
-            print(f"Invalid value: {user_value}. Please enter an integer.")
-            return get_user_value(field_, value, provider, cdf_cluster, cdf_project, idp_tenant_id)
-    return user_value
-
-
-def _is_unchanged(
-    current: EnvironmentVariables, provider: Provider, login_flow: LoginFlow, cdf_project: str, cdf_cluster: str
-) -> bool:
-    return (
-        current.PROVIDER == provider
-        and current.LOGIN_FLOW == login_flow
-        and current.CDF_PROJECT == cdf_project
-        and current.CDF_CLUSTER == cdf_cluster
-    )
-
-
-if __name__ == "__main__":
-    # For easy testing
-    envs = prompt_user_environment_variables()
-    print(envs)
