+++ cognite_toolkit/_cdf_tk/commands/auth/data_classes/_types.py
@@ -1,0 +1,4 @@
+from typing import Literal, TypeAlias
+
+Provider: TypeAlias = Literal["entra_id", "auth0", "cdf", "other"]
+LoginFlow: TypeAlias = Literal["client_credentials", "token", "device_code", "interactive", "session"]
