+++ cognite_toolkit/_cdf_tk/commands/auth/session_command.py
@@ -6,9 +6,9 @@ from rich import print
 
 from cognite_toolkit._cdf_tk.commands._base import ToolkitCommand
 from cognite_toolkit._cdf_tk.exceptions import AuthenticationError
-from cognite_toolkit._cdf_tk.utils.auth import LoginFlow, read_env_login_flow
 
 from .cogidp import SessionProject, fetch_session_user_info
+from .data_classes import EnvironmentVariables, LoginFlow
 from .oidc import login_for_session, revoke_refresh_token
 from .session_keyring import read_session_token
 from .session_refresh import SessionExpiredError, ensure_fresh_session
@@ -23,7 +23,7 @@ from .session_store import (
 
 def confirm_login_flow_overrides_env(selected_flow: LoginFlow) -> bool:
     """Warn and confirm when login uses a different mode than .env."""
-    env_flow = read_env_login_flow()
+    env_flow = EnvironmentVariables.login_flow_from_environment()
     if env_flow is None or env_flow == selected_flow:
         return True
