+++ tests/test_unit/test_cdf_tk/test_commands/test_run.py
@@ -18,9 +18,9 @@ from cognite_toolkit._cdf_tk.commands import (
     RunTransformationCommand,
     RunWorkflowCommand,
 )
+from cognite_toolkit._cdf_tk.commands.auth import EnvironmentVariables
 from cognite_toolkit._cdf_tk.commands.build_v2.data_classes import BuildLineage
 from cognite_toolkit._cdf_tk.commands.run import FunctionCallArgs
-from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.data import RUN_DATA
 from tests.test_unit.approval_client import ApprovalToolkitClient
