+++ tests/test_unit/test_cdf_tk/test_cruds/test_data_model.py
@@ -23,11 +23,11 @@ from cognite_toolkit._cdf_tk.client.resource_classes.graphql_data_model import (
     GraphQLDataModelResponse,
 )
 from cognite_toolkit._cdf_tk.client.testing import monkeypatch_toolkit_client
+from cognite_toolkit._cdf_tk.commands.auth import EnvironmentVariables
 from cognite_toolkit._cdf_tk.constants import VIEW_UPSERT_BATCH_LIMIT
 from cognite_toolkit._cdf_tk.exceptions import ToolkitCycleError
 from cognite_toolkit._cdf_tk.resource_ios import DataModelIO, EdgeCRUD, NodeCRUD, SpaceCRUD
 from cognite_toolkit._cdf_tk.resource_ios._resource_ios import GraphQLCRUD, ViewIO
-from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.test_unit.approval_client import ApprovalToolkitClient
 from tests.utils import to_deploy_status
