+++ tests/test_unit/test_cdf_tk/test_cruds/test_extraction_pipeline.py
@@ -13,6 +13,7 @@ from cognite_toolkit._cdf_tk.client.resource_classes.extraction_pipeline_config 
     ExtractionPipelineConfigRequest,
     ExtractionPipelineConfigResponse,
 )
+from cognite_toolkit._cdf_tk.commands.auth import EnvironmentVariables
 from cognite_toolkit._cdf_tk.resource_ios import (
     DataSetsIO,
     ExtractionPipelineConfigIO,
@@ -21,7 +22,6 @@ from cognite_toolkit._cdf_tk.resource_ios import (
     RawTableCRUD,
     ResourceIO,
 )
-from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.test_unit.approval_client import ApprovalToolkitClient
 from tests.utils import to_deploy_status
