+++ tests/test_unit/test_cdf_tk/test_cruds/test_function.py
@@ -26,11 +26,11 @@ from cognite_toolkit._cdf_tk.client.resource_classes.function_schedule import (
 )
 from cognite_toolkit._cdf_tk.client.testing import monkeypatch_toolkit_client
 from cognite_toolkit._cdf_tk.commands import DeployV2Command
+from cognite_toolkit._cdf_tk.commands.auth import EnvironmentVariables
 from cognite_toolkit._cdf_tk.commands.deploy_v2.command import ReadResource
 from cognite_toolkit._cdf_tk.exceptions import ResourceCreationError, ToolkitRequiredValueError
 from cognite_toolkit._cdf_tk.resource_ios import FunctionIO, FunctionScheduleIO
 from cognite_toolkit._cdf_tk.utils import calculate_directory_hash, calculate_secure_hash
-from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.data import LOAD_DATA
 from tests.test_unit.approval_client import ApprovalToolkitClient
